+++ R/vis_clus.R
@@ -17,6 +17,8 @@
 #' [ggplotly()][plotly::ggplotly] then it's best to set this to `FALSE`.
 #' @param image_id A `character(1)` with the name of the image ID you want to
 #' use in the background.
+#' @param alpha A `numeric(1)` in the `[0, 1]` range that specifies the
+#' transparency level of the data on the spots.
 #' @param ... Passed to [paste0()][base::paste] for making the title of the
 #' plot following the `sampleid`.
 #'
@@ -73,6 +75,7 @@ vis_clus <- function(spe,
     ),
     spatial = TRUE,
     image_id = "lowres",
+    alpha = 1,
     ...) {
     spe_sub <- spe[, spe$sample_id == sampleid]
     d <- as.data.frame(colData(spe_sub, spatialData = TRUE, spatialCoords = TRUE), optional = TRUE)
@@ -85,6 +88,7 @@ vis_clus <- function(spe,
         spatial = spatial,
         title = paste0(sampleid, ...),
         colors = get_colors(colors, d[, clustervar]),
-        image_id = image_id
+        image_id = image_id,
+        alpha = alpha
     )
 }